db: read clustering
